Flytt converters for vedtaksstatistikk til egen fil
